pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0"
description = ""
authors = ["notsoyoung2 <p.sergey1997@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8.1"
Django = "^4.1.5"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
django-filter = "^22.1"
python-gettext = "^4.1"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
